modified:   framework/api.c 	modified:   framework/api.h 	modified:   framework/client.c 	modified:   framework/types.h 	modified:   framework/worker.c

+++ framework/api.h
@@ -9,7 +9,7 @@ void api_recv_free(struct api_msg *msg);
 void api_state_free(struct api_state *state);
 void api_state_init(struct api_state *state, int fd);
 
-int api_send(struct api_state *state, struct api_msg *msg);
+ssize_t api_send(struct api_state *state, struct api_msg *msg);
 void api_send_free(struct api_msg *msg);
 
 /* TODO add API calls to send messages to perform client-server interactions */
